restructure api code
