+++ api/utils/generateSetOfNullifiedFields.ts
