--- pages/api/users/index.ts
@@ -1,27 +1,0 @@
-import { NextApiRequest, NextApiResponse } from "next";
-
-export type User = {
-  id: number;
-  name: string;
-};
-
-export const sampleUserData: User[] = [
-  { id: 101, name: "Alice" },
-  { id: 102, name: "Bob" },
-  { id: 103, name: "Caroline" },
-  { id: 104, name: "Dave" },
-];
-
-const handler = (_req: NextApiRequest, res: NextApiResponse) => {
-  try {
-    if (!Array.isArray(sampleUserData)) {
-      throw new Error("Cannot find user data");
-    }
-
-    res.status(200).json(sampleUserData);
-  } catch (err) {
-    res.status(500).json({ statusCode: 500, message: err.message });
-  }
-};
-
-export default handler;
